+++ src/lib.rs
@@ -66,7 +66,6 @@ pub mod utils;
 #[doc(inline)]
 pub use teloxide_core::*;
 
-use teloxide_core::requests::ResponseResult;
 #[cfg(feature = "macros")]
 // FIXME(waffle): use `docsrs` here when issue with combine is resolved <https://github.com/teloxide/teloxide/pull/305#issuecomment-716172103>
 #[cfg_attr(all(teloxide_docsrs, feature = "nightly"), doc(cfg(feature = "macros")))]
@@ -75,6 +74,8 @@ pub use teloxide_macros::teloxide;
 #[cfg(all(feature = "nightly", doctest))]
 #[doc(include = "../README.md")]
 enum ReadmeDocTests {}
+
+use teloxide_core::requests::ResponseResult;
 
 /// A shortcut for `ResponseResult::Ok(val)`.
 pub fn respond<T>(val: T) -> ResponseResult<T> {